apt: add functionality to query packages by origin
